@@ -14,13 +14,14 @@ description: 4chan viewer written in Flutter
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
-  flutter: ">=1.10.0 <3.0.0"
+  sdk: '>=3.0.0 <4.0.0'
+  flutter: ">=1.10.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
+  json_theme: ^5.0.2+6
   collection: ^1.17.0
 
   flutter_bloc: ^8.1.2
@@ -65,7 +66,8 @@ dependencies:
   video_thumbnail: ^0.5.3
   dart_vlc: ^0.4.0
   carousel_slider: ^4.2.1
-  sliding_sheet: ^0.5.2 # TODO - replace with https://pub.dev/packages/sliding_up_panel
+#  sliding_sheet: ^0.5.2 # TODO - replace with https://pub.dev/packages/sliding_up_panel
+  sliding_up_panel: ^2.0.0+1
 
 dev_dependencies:
   drift_dev: ^2.7.0
@@ -93,6 +95,10 @@ flutter:
   # assets:
   #  - images/a_dot_burr.jpeg
   #  - images/a_dot_ham.jpeg
+
+  assets:
+    - assets/appainter_theme_light.json
+    - assets/appainter_theme_dark.json
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
